Description: The static version evaluation tool scans for non-compliant code in lightweight UI components IssueNo: https://gitee.com/openharmony/arkui_ui_lite/issues/I7EHKK Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: guanzengkun <guanzengkun@huawei.com>

+++ frameworks/font/glyphs_file.cpp
@@ -165,7 +165,7 @@ int8_t GlyphsFile::SetFile(const char* fontName, int32_t fp, uint32_t start)
     }
 
     ret = read(fp_, fontHeaderCache_, size);
-    if (ret != size) {
+    if (ret != static_cast<int32_t>(size)) {
         GRAPHIC_LOGE("GlyphsFile::SetFile read failed");
         return INVALID_RET_VALUE;
     }
